refactor: Get rid of asynchronous eventing in test
