+++ test/integration/monitor_tampered_test.go
@@ -1,9 +1,9 @@
 package integration
 
 import (
+	"bytes"
 	"context"
 	"testing"
-	"time"
 )
 
 // tamperCheckpointRootHash modifies a checkpoint file by appending "tampered" to its root hash
@@ -21,19 +21,29 @@ func tamperCheckpointRootHash(t *testing.T, checkpointFile string) {
 
 func TestTamperedCheckpoint(t *testing.T) {
 	mockServer := RekorServer().WithData().Build()
-	binary, checkpointFile, monitorPort := setupTest(t, mockServer.URL, false)
 	defer mockServer.Close()
 
-	ctx, cancel := context.WithTimeout(context.Background(), 30*time.Second)
-	defer cancel()
-
+	ctx, cancel := context.WithCancel(context.Background())
+	checkpointFile := createCheckpointFile(ctx, t, mockServer.URL, true)
 	t.Run("validate_and_tamper_checkpoint_file", func(t *testing.T) {
 		tamperCheckpointRootHash(t, checkpointFile)
 	})
 
-	runCmd, logs := startMonitor(t, ctx, binary, checkpointFile, monitorPort, mockServer.URL)
+	monitorPort, err := findFreePort()
+	if err != nil {
+		t.Fatalf("failed to find free port: %v", err)
+	}
+	runCmd := startMonitorCommand(ctx, checkpointFile, monitorPort, mockServer.URL)
+	logs := bytes.NewBuffer(nil)
+	runCmd.Stderr = logs
+	if err := runCmd.Start(); err != nil {
+		t.Fatalf("failed to start monitor: %v", err)
+	}
 
-	metrics := fetchMetrics(t, monitorPort)
+	metrics, err := fetchMetrics(monitorPort)
+	if err != nil {
+		t.Fatalf("failed to fetch metrics: %v", err)
+	}
 	validateLogsAndMetrics(t, logs, metrics, MonitorExpectations{
 		ExpectErrorLog:       true,
 		ExpectedErrorType:    "error running consistency check",
@@ -41,5 +51,7 @@ func TestTamperedCheckpoint(t *testing.T) {
 		ExpectedTotalCount:   1,
 	})
 
-	stopMonitor(t, cancel, runCmd)
+	cancel()
+	// Wait for the monitor to exit, test timeouts if it doesn't
+	runCmd.Wait()
 }
